requirements.txt
datasets==2.21.0
fvcore==0.1.5.post20221221
matplotlib==3.8.3
numpy==1.25.1
pandas==2.2.2
peft==0.10.0
PyYAML==6.0.1
scikit_learn==1.4.0
scikit_surprise==1.1.3
scipy==1.14.0
torch==2.4.0+cu121
tqdm==4.65.0
transformers==4.43.4

#vllm==0.5.4+cu121
#deepspeed
#accelerate==0.27.1
